+++ src/plugins/installed-plugin-index-store.test.ts
@@ -438,6 +438,11 @@ describe("installed plugin index persistence", () => {
             clawhubPackage: "clawpack-demo",
             clawhubFamily: "code-plugin",
             clawhubChannel: "official",
+            artifactKind: "npm-pack",
+            artifactFormat: "tgz",
+            npmIntegrity: "sha512-clawpack",
+            npmShasum: "1".repeat(40),
+            npmTarballName: "clawpack-demo-2026.5.1-beta.2.tgz",
             clawpackSha256: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
             clawpackSpecVersion: 1,
             clawpackManifestSha256:
@@ -474,6 +479,11 @@ describe("installed plugin index persistence", () => {
           clawhubPackage: "clawpack-demo",
           clawhubFamily: "code-plugin",
           clawhubChannel: "official",
+          artifactKind: "npm-pack",
+          artifactFormat: "tgz",
+          npmIntegrity: "sha512-clawpack",
+          npmShasum: "1".repeat(40),
+          npmTarballName: "clawpack-demo-2026.5.1-beta.2.tgz",
           clawpackSha256: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           clawpackSpecVersion: 1,
           clawpackManifestSha256: